requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
-e file:.
    # via -r requirements.in
backports.entry-points-selectable==1.1.0
    # via virtualenv
black==21.9b0
    # via -r requirements.in
bleach==4.1.0
    # via readme-renderer
certifi==2021.10.8
    # via requests
cffi==1.14.6
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   black
    #   pip-tools
colorama==0.4.4
    # via twine
coverage==6.0.2
    # via -r requirements.in
cryptography==35.0.0
    # via secretstorage
distlib==0.3.3
    # via virtualenv
docutils==0.17.1
    # via readme-renderer
filelock==3.3.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
identify==2.3.0
    # via pre-commit
idna==3.2
    # via requests
importlib-metadata==4.8.1
    # via
    #   keyring
    #   twine
isort==5.9.3
    # via -r requirements.in
jeepney==0.7.1
    # via
    #   keyring
    #   secretstorage
keyring==23.2.1
    # via twine
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.0
    # via bleach
pathspec==0.9.0
    # via black
pep517==0.11.0
    # via pip-tools
pip-tools==6.3.1
    # via -r requirements.in
pkginfo==1.7.1
    # via twine
platformdirs==2.4.0
    # via
    #   black
    #   virtualenv
pre-commit==2.15.0
    # via -r requirements.in
pycodestyle==2.8.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.4.0
    # via flake8
pygments==2.10.0
    # via readme-renderer
pyparsing==2.4.7
    # via packaging
python-magic==0.4.24
    # via querido-diario-toolbox
pyyaml==5.4.1
    # via pre-commit
readme-renderer==30.0
    # via twine
regex==2021.10.8
    # via black
requests==2.26.0
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==0.9.1
    # via twine
rfc3986==1.5.0
    # via twine
secretstorage==3.3.1
    # via keyring
six==1.16.0
    # via
    #   bleach
    #   virtualenv
toml==0.10.2
    # via pre-commit
tomli==1.2.1
    # via
    #   black
    #   pep517
tqdm==4.62.3
    # via twine
twine==3.4.2
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via black
urllib3==1.26.7
    # via requests
virtualenv==20.8.1
    # via pre-commit
webencodings==0.5.1
    # via bleach
wheel==0.37.0
    # via pip-tools
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
